# Trie changes