--- a/runtime/storage/src/trie/mod.rs
+++ b/runtime/storage/src/trie/mod.rs
@@ -4,7 +4,6 @@ use byteorder::{LittleEndian, ReadBytesExt, WriteBytesExt};
 use cached::{Cached, SizedCache};
 pub use kvdb::DBValue;
 use primitives::hash::{hash, CryptoHash};
-use std::collections::hash_map::Entry;
 use std::collections::HashMap;
 use std::convert::TryFrom;
 use std::fmt;