- change structure, make an storage object for temporary mutable trie nodes
- normalize nodes after delete, a test for unique representation
- non-recursive trie methods, an evil contract test
- correctly compute refcounts
- TrieChanges to support storing multiple versions in the future
- Cache for storage, will be more useful later, also fix cached crate

Performance got better but still a lot of time is spent allocating.