@mikhailOK 2019-05-07
master:
```
test runtime_send_money                     ... bench:     170,699 ns/iter (+/- 11,828)
test runtime_wasm_bad_code                  ... bench:     494,058 ns/iter (+/- 18,055)
test runtime_wasm_benchmark_10_reads_legacy ... bench:     650,622 ns/iter (+/- 68,536)
test runtime_wasm_benchmark_storage_100     ... bench:   1,427,961 ns/iter (+/- 69,918)
test runtime_wasm_benchmark_storage_1000    ... bench:  10,746,114 ns/iter (+/- 1,372,860)
test runtime_wasm_benchmark_sum_1000        ... bench:     522,782 ns/iter (+/- 33,496)
test runtime_wasm_benchmark_sum_1000000     ... bench:  12,837,735 ns/iter (+/- 706,853)
test runtime_wasm_set_value                 ... bench:     514,188 ns/iter (+/- 68,692)
```
this change:
```
test runtime_send_money                     ... bench:     171,727 ns/iter (+/- 32,866)
test runtime_wasm_bad_code                  ... bench:     539,352 ns/iter (+/- 84,401)
test runtime_wasm_benchmark_10_reads_legacy ... bench:     632,613 ns/iter (+/- 201,592)
test runtime_wasm_benchmark_storage_100     ... bench:   1,126,283 ns/iter (+/- 88,090)
test runtime_wasm_benchmark_storage_1000    ... bench:   7,214,292 ns/iter (+/- 615,960)
test runtime_wasm_benchmark_sum_1000        ... bench:     541,049 ns/iter (+/- 138,420)
test runtime_wasm_benchmark_sum_1000000     ... bench:  13,147,020 ns/iter (+/- 1,248,270)
test runtime_wasm_set_value                 ... bench:     533,525 ns/iter (+/- 110,945)
```
note that cache actually slows things down right now because we clear it, that will be fixed later